Update java package names
