+++ compiler-plugin/pom.xml
@@ -29,6 +29,7 @@
     <modelVersion>4.0.0</modelVersion>
     <artifactId>jms-compiler-plugin</artifactId>
     <version>0.6.0</version>
+    <packaging>jar</packaging>
 
     <dependencies>
         <dependency>
