+++ compiler-plugin/src/main/resources/META-INF/services/org.ballerinalang.compiler.plugins.CompilerPlugin
@@ -1,1 +1,1 @@
-org.wso2.ei.module.jms.ServiceCompilerPlugin
+org.wso2.ei.b7a.jms.ServiceCompilerPlugin
