+++ utils/src/main/java/org/wso2/ei/b7a/jms/BallerinaJmsException.java
@@ -17,31 +17,18 @@
  *
  */
 
-package org.wso2.ei.module.jms;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+package org.wso2.ei.b7a.jms;
 
 /**
- * Utility class for JMS related common operations.
+ * Representation of the custom exception in the JMS module
  */
-public class JmsUtils {
+public class BallerinaJmsException extends Exception {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(JmsUtils.class);
-
-    /**
-     * Utility class cannot be instantiated.
-     */
-    private JmsUtils() {
+    BallerinaJmsException(String message) {
+        super(message);
     }
 
-    /**
-     * Check given string is not null or empty after trimming
-     * @param str String value
-     * @return true/false based on the input
-     */
-    static boolean notNullOrEmptyAfterTrim(String str) {
-        return !(str == null || str.trim().isEmpty());
+    public BallerinaJmsException(String message, Throwable cause) {
+        super(message, cause);
     }
 }
-
