+++ utils/src/main/java/org/wso2/ei/b7a/jms/LoggingExceptionListener.java
@@ -17,7 +17,7 @@
  *
  */
 
-package org.wso2.ei.module.jms;
+package org.wso2.ei.b7a.jms;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
